ux: Add missing Preact error boundary to the dashboard
Closes #94
